requirements.txt
acres
aiofiles
aiohappyeyeballs
aiohttp
aiosignal
annotated-types
anyio
async-timeout
attrs
blinker
cachetools
certifi
cffi
charset-normalizer
ci-info
click
colorama
configobj
configparser
cryptography
dataclasses-json
etelemetry
exceptiongroup
faiss-cpu
filelock
filetype
flask
flask-sqlalchemy
frozenlist
fsspec
google-ai-generativelanguage
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-generativeai
googleapis-common-protos
greenlet
grpcio
grpcio-status
h11
httpcore
httplib2
httpx
httpx-sse
huggingface-hub
idna
importlib-resources
isodate
itsdangerous
jinja2
joblib
jsonify
jsonpatch
jsonpointer
langchain
langchain-community
langchain-core
langchain-google-genai
langchain-text-splitters
langsmith
looseversion
lxml
MarkupSafe
marshmallow
mpmath
multidict
mypy-extensions
networkx
nibabel
nipype
numpy
orjson
packaging
pandas
pathlib
pdfminer.six
pillow
propcache
proto-plus
protobuf
prov
psycopg
psycopg-binary
psycopg-pool
psycopg2
puremagic
pyasn1
pyasn1-modules
pycparser
pydantic
pydantic-core
pydantic-settings
pydot
PyMuPDFb
pyparsing
python-dateutil
python-dotenv
pytz
pyxnat
PyYAML
rdflib
regex
requests
requests-toolbelt
rsa
safetensors
scikit-learn
scipy
sentence-transformers
simplejson
six
sniffio
SQLAlchemy
sqlparse
starlette
sympy
tenacity
threadpoolctl
tokenizers
torch
tqdm
traits
transformers
typing-extensions
typing-inspect
tzdata
uritemplate
urllib3
uvicorn
waitress
werkzeug
yarl
zstandard